Path操作をオブジェクトに取りまとめた
