+++ src/main/scala-2.11/plantuml/PlantUmlImageGenerator.scala
@@ -1,7 +1,5 @@
 package plantuml
 
-import java.io.File
-
 import ammonite.ops._
 import net.sourceforge.plantuml.SourceFileReader
 
@@ -9,6 +7,8 @@ import net.sourceforge.plantuml.SourceFileReader
  * Created by shinsuke-abe on 2015/10/02.
  */
 object PlantUmlImageGenerator {
+  import common.PathUtils._
+
   def apply(inputFile: Path, outputDir: Path): Either[Throwable, Path] = {
     require(exists! inputFile && (stat! inputFile).isFile)
 
@@ -21,9 +21,5 @@ object PlantUmlImageGenerator {
       if (generatedImage.lineErrorRaw() > 0) Left(new Exception(generatedImage.getDescription))
       else Right(Path(generatedImage.getPngFile.toString))
     }
-  }
-
-  implicit class RichPath(path: Path) {
-    def toFile = new File(path.toString())
   }
 }
